Changed default int type to long. Added refs to shared libraries to fix issue with .net libs not finding it.

+++ C/include/libCpScript.Asm.h
@@ -57,12 +57,12 @@ EXPORT void State_PushBool(void* State, short v);
 EXPORT void State_PushString(void* State, char* v);
 EXPORT void State_Push(void* State, void* v);
 
-EXPORT int State_GetIntVariableInScope(void* S, char* n);
+EXPORT long State_GetIntVariableInScope(void* S, char* n);
 EXPORT double State_GetDoubleVariableInScope(void* S, char* n);
 EXPORT short State_GetBoolVariableInScope(void* S, char* n);
 EXPORT char* State_GetStringVariableInScope(void* S, char* n);
 
-EXPORT void State_SetIntVariableInScope(void* S, char* n, int v);
+EXPORT void State_SetIntVariableInScope(void* S, char* n, long v);
 EXPORT void State_SetDoubleVariableInScope(void* S, char* n, double v);
 EXPORT void State_SetBoolVariableInScope(void* S, char* n, short v);
 EXPORT void State_SetStringVariableInScope(void* S, char* n, char* v);
